Integrate SNAPHU in InSAR workflow (#968)
* Move correlation_thresh_increments to common ICU, Phass parameters

* Add snaphu parameters in InSAR schema

* Re-organize unwrap.py to include snaphu unwrapper

* Integrate SNAPHU in InSAR workflow

* Correct rg bandwidth, conn components to uint32, fix stats

* Add functionality to clip nan from correlation data

* add support for unsigned int rasters to CUDA geocode

* Fix computation azimuth resolution. Log ICU and Phass attributes

* Correct typo in clean cfg function

* Rename some functionalities for clarity

* Expose and include MCF initializer

* add raster data type and interpolation method check in CUDA geocode

* Remove default value comments for SNAPHU in InSAR schema

* Remove correlation and phase std model from schema and unwrap.py

* Fix identation doc strings

* Change Snaphu schema params name to match Python wrapper

* Use ternary if conditions and fix params in schema

* Resolve param name diff between Python interface and InSAR schema

* limit int type interp to nearest neighbor

* Move unwrap default values to InSAR default

* Fix pep8 notation mistakes

* remove erroneously introduced file

* fix incorrect init logic

* added int type and interp type check
whitespace cleanup

* Include SNAPHU in Rosamond insar test case

* Remove function cleaning NaNs

* Correct bug on band from where to read data

* Move unwrap-related comments from schema to InSAR default runconfig

* Correct get function for cfg info extraction

* Fix bug of connected component type

+++ python/packages/nisar/workflows/h5_prep.py
@@ -746,7 +746,7 @@ def prep_ds_insar(pcfg, dst, dst_h5):
                                         units="radians", grids=grids_val,
                                         long_name='unwrapped phase')
                       descr = f"Connected components for {pol} layer"
-                      _create_datasets(dst_h5[pol_path], igram_shape, np.uint8,
+                      _create_datasets(dst_h5[pol_path], igram_shape, np.uint32,
                                       'connectedComponents', descr=descr, units=" ",
                                       grids=grids_val,
                                       long_name='connected components')
